add cancel function

+++ app/src/main/java/com/wirelessalien/zipxtract/service/ExtractArchiveService.kt
@@ -22,7 +22,10 @@ import android.app.NotificationChannel
 import android.app.NotificationManager
 import android.app.PendingIntent
 import android.app.Service
+import android.content.BroadcastReceiver
+import android.content.Context
 import android.content.Intent
+import android.content.IntentFilter
 import android.media.MediaScannerConnection
 import android.os.Build
 import android.os.Environment
@@ -35,6 +38,7 @@ import androidx.localbroadcastmanager.content.LocalBroadcastManager
 import androidx.preference.PreferenceManager
 import com.wirelessalien.zipxtract.R
 import com.wirelessalien.zipxtract.activity.MainActivity
+import com.wirelessalien.zipxtract.constant.BroadcastConstants.ACTION_CANCEL_OPERATION
 import com.wirelessalien.zipxtract.constant.BroadcastConstants.ACTION_EXTRACTION_COMPLETE
 import com.wirelessalien.zipxtract.constant.BroadcastConstants.ACTION_EXTRACTION_ERROR
 import com.wirelessalien.zipxtract.constant.BroadcastConstants.ACTION_EXTRACTION_PROGRESS
@@ -94,6 +98,17 @@ class ExtractArchiveService : Service() {
 
     private var archiveFormat: ArchiveFormat? = null
     private var extractionJob: Job? = null
+    private var progressMonitor: ProgressMonitor? = null
+
+    private val cancelReceiver = object : BroadcastReceiver() {
+        override fun onReceive(context: Context?, intent: Intent?) {
+            if (intent?.action == ACTION_CANCEL_OPERATION) {
+                progressMonitor?.isCancelAllTasks = true
+                extractionJob?.cancel()
+                stopForegroundService()
+            }
+        }
+    }
 
     override fun onBind(intent: Intent): IBinder? = null
 
@@ -101,6 +116,7 @@ class ExtractArchiveService : Service() {
         super.onCreate()
         fileOperationsDao = FileOperationsDao(this)
         createNotificationChannel()
+        LocalBroadcastManager.getInstance(this).registerReceiver(cancelReceiver, IntentFilter(ACTION_CANCEL_OPERATION))
     }
 
     override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
@@ -137,6 +153,7 @@ class ExtractArchiveService : Service() {
     override fun onDestroy() {
         super.onDestroy()
         extractionJob?.cancel()
+        LocalBroadcastManager.getInstance(this).unregisterReceiver(cancelReceiver)
     }
 
     private fun createNotificationChannel() {
@@ -244,16 +261,24 @@ class ExtractArchiveService : Service() {
                         sendLocalBroadcast(Intent(ACTION_EXTRACTION_COMPLETE).putExtra(EXTRA_DIR_PATH, destinationDir.absolutePath))
                     }
                 } catch (e: SevenZipException) {
-                    e.printStackTrace()
-                    showErrorNotification(e.message ?: getString(R.string.general_error_msg))
-                    sendLocalBroadcast(Intent(ACTION_EXTRACTION_ERROR).putExtra(EXTRA_ERROR_MESSAGE, e.message ?: getString(R.string.general_error_msg)))
+                    if (e.message == "Cancelled") {
+                        // Cancelled by user, do nothing
+                    } else {
+                        e.printStackTrace()
+                        showErrorNotification(e.message ?: getString(R.string.general_error_msg))
+                        sendLocalBroadcast(Intent(ACTION_EXTRACTION_ERROR).putExtra(EXTRA_ERROR_MESSAGE, e.message ?: getString(R.string.general_error_msg)))
+                    }
                 } finally {
                     inArchive.close()
                 }
             } catch (e: SevenZipException) {
-                e.printStackTrace()
-                showErrorNotification(e.message ?: getString(R.string.general_error_msg))
-                sendLocalBroadcast(Intent(ACTION_EXTRACTION_ERROR).putExtra(EXTRA_ERROR_MESSAGE, e.message ?: getString(R.string.general_error_msg)))
+                if (e.message == "Cancelled") {
+                    // Cancelled by user, do nothing
+                } else {
+                    e.printStackTrace()
+                    showErrorNotification(e.message ?: getString(R.string.general_error_msg))
+                    sendLocalBroadcast(Intent(ACTION_EXTRACTION_ERROR).putExtra(EXTRA_ERROR_MESSAGE, e.message ?: getString(R.string.general_error_msg)))
+                }
             } finally {
                 inStream.close()
                 if (useAppNameDir) {
@@ -297,6 +322,9 @@ class ExtractArchiveService : Service() {
                                 ): ByteBuffer? {
                                     buffer.clear()
                                     try {
+                                        if (extractionJob?.isActive == false) {
+                                            throw IOException("Cancelled")
+                                        }
                                         val bytesRead = Os.read(clientData, buffer)
                                         bytesProcessed += bytesRead
                                         val progress = (bytesProcessed * 100 / totalBytes).toInt()
@@ -389,15 +417,19 @@ class ExtractArchiveService : Service() {
                                 .putExtra(EXTRA_DIR_PATH, destinationDir.absolutePath)
                         )
                     } catch (e: Exception) {
-                        e.printStackTrace()
-                        val errorMessage = e.message ?: getString(R.string.general_error_msg)
-                        showErrorNotification(errorMessage)
-                        sendLocalBroadcast(
-                            Intent(ACTION_EXTRACTION_ERROR).putExtra(
-                                EXTRA_ERROR_MESSAGE,
-                                errorMessage
+                        if (e.message == "Cancelled") {
+                            // Cancelled
+                        } else {
+                            e.printStackTrace()
+                            val errorMessage = e.message ?: getString(R.string.general_error_msg)
+                            showErrorNotification(errorMessage)
+                            sendLocalBroadcast(
+                                Intent(ACTION_EXTRACTION_ERROR).putExtra(
+                                    EXTRA_ERROR_MESSAGE,
+                                    errorMessage
+                                )
                             )
-                        )
+                        }
                     } finally {
                         if (archive != 0L) {
                             Archive.free(archive)
@@ -484,6 +516,9 @@ class ExtractArchiveService : Service() {
                         FileOutputStream(outputFile).use { output ->
                             var n: Int
                             while (tarInput.read(buffer).also { n = it } != -1) {
+                                    if (extractionJob?.isActive == false) {
+                                        return
+                                    }
                                 output.write(buffer, 0, n)
                                 bytesRead += n
                                 val progress = (bytesRead * 100 / totalBytes).toInt()
@@ -501,10 +536,15 @@ class ExtractArchiveService : Service() {
             showCompletionNotification(destinationDir.absolutePath)
             sendLocalBroadcast(Intent(ACTION_EXTRACTION_COMPLETE).putExtra(EXTRA_DIR_PATH, destinationDir.absolutePath))
         } catch (e: IOException) {
-            e.printStackTrace()
-            tryLibArchiveAndroid(file, destinationDir)
-            showErrorNotification(e.message ?: getString(R.string.general_error_msg))
-            sendLocalBroadcast(Intent(ACTION_EXTRACTION_ERROR).putExtra(EXTRA_ERROR_MESSAGE, e.message ?: getString(R.string.general_error_msg)))
+            if (e.message == "Cancelled") {
+                // Cancelled
+            } else {
+                e.printStackTrace()
+                tryLibArchiveAndroid(file, destinationDir)
+                if (extractionJob?.isActive == false) return
+                showErrorNotification(e.message ?: getString(R.string.general_error_msg))
+                sendLocalBroadcast(Intent(ACTION_EXTRACTION_ERROR).putExtra(EXTRA_ERROR_MESSAGE, e.message ?: getString(R.string.general_error_msg)))
+            }
         }
     }
 
@@ -567,22 +607,27 @@ class ExtractArchiveService : Service() {
             }
             zipFile.extractAll(destinationDir.absolutePath)
 
-            val progressMonitor = zipFile.progressMonitor
-            while (!progressMonitor.state.equals(ProgressMonitor.State.READY)) {
-                if (progressMonitor.state.equals(ProgressMonitor.State.BUSY)) {
-                    val percentDone = (progressMonitor.percentDone)
+            progressMonitor = zipFile.progressMonitor
+            while (!progressMonitor!!.state.equals(ProgressMonitor.State.READY)) {
+                if (progressMonitor!!.state.equals(ProgressMonitor.State.BUSY)) {
+                    val percentDone = (progressMonitor!!.percentDone)
                     startForeground(NOTIFICATION_ID, createNotification(percentDone))
                     sendLocalBroadcast(Intent(ACTION_EXTRACTION_PROGRESS).putExtra(EXTRA_PROGRESS, percentDone))
                 }
                 Thread.sleep(100)
             }
-            FileUtils.setLastModifiedTime(directories)
-            scanForNewFiles(destinationDir)
-            showCompletionNotification(destinationDir.absolutePath)
-            sendLocalBroadcast(Intent(ACTION_EXTRACTION_COMPLETE).putExtra(EXTRA_DIR_PATH, destinationDir.absolutePath))
 
-            if (useAppNameDir) {
-                filesDir.deleteRecursively()
+            if (progressMonitor!!.result == ProgressMonitor.Result.CANCELLED) {
+                 // Do nothing
+            } else {
+                FileUtils.setLastModifiedTime(directories)
+                scanForNewFiles(destinationDir)
+                showCompletionNotification(destinationDir.absolutePath)
+                sendLocalBroadcast(Intent(ACTION_EXTRACTION_COMPLETE).putExtra(EXTRA_DIR_PATH, destinationDir.absolutePath))
+
+                if (useAppNameDir) {
+                    filesDir.deleteRecursively()
+                }
             }
 
         } catch (e: ZipException) {
@@ -734,6 +779,9 @@ class ExtractArchiveService : Service() {
         override fun prepareOperation(p0: ExtractAskMode?) {}
 
         override fun setCompleted(complete: Long) {
+            if (extractionJob?.isActive == false) {
+                throw SevenZipException("Cancelled")
+            }
             val progress = ((complete.toDouble() / totalSize) * 100).toInt()
             startForeground(NOTIFICATION_ID, createNotification(progress))
             updateProgress(progress)
